backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
PyYAML==6.0.2
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
pytest==8.3.3
